Moved item grab menu patch
